feat : 일기태그목록 조회시 해당태그에 대한 일기갯수 필드 추가

+++ server/src/main/java/com/codemouse/salog/tags/diaryTags/repository/DiaryTagLinkRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 public interface DiaryTagLinkRepository extends JpaRepository<DiaryTagLink, Long> {
     List<DiaryTagLink> findByDiaryTagTagNameAndDiaryTagMember(String tagName, Member member);
-    Long countByDiaryTag(DiaryTag diaryTag);
+    int countByDiaryTag(DiaryTag diaryTag);
 
     @Transactional
     @Modifying // update나 delete는 해당 어노테이션을 추가해줘야함. (@Query만 사용할 경우 기본 select)
